Refactor parametrized tests (#99)
