+++ tests/test_LocalClassifierPerParentNode.py
@@ -8,12 +8,10 @@ from numpy.testing import assert_array_equal
 from scipy.sparse import csr_matrix
 from sklearn.exceptions import NotFittedError
 from sklearn.linear_model import LogisticRegression
-from sklearn.neighbors import KNeighborsClassifier
 from sklearn.utils.estimator_checks import parametrize_with_checks
 from sklearn.utils.validation import check_is_fitted
 
 from hiclass import LocalClassifierPerParentNode
-from hiclass.ConstantClassifier import ConstantClassifier
 
 
 @parametrize_with_checks([LocalClassifierPerParentNode()])
@@ -86,18 +84,6 @@ def test_fit_digraph_joblib_multiprocessing(digraph_logistic_regression):
                 digraph_logistic_regression.hierarchy_.nodes[node]["classifier"]
             )
     assert 1
-
-
-def test_fit_1_class():
-    lcppn = LocalClassifierPerParentNode(
-        local_classifier=LogisticRegression(), n_jobs=2
-    )
-    y = np.array([["1", "2"]])
-    X = np.array([[1, 2]])
-    ground_truth = np.array([["1", "2"]])
-    lcppn.fit(X, y)
-    prediction = lcppn.predict(X)
-    assert_array_equal(ground_truth, prediction)
 
 
 @pytest.fixture
@@ -198,89 +184,3 @@ def test_fit_predict():
     lcppn.fit(x, y)
     predictions = lcppn.predict(x)
     assert_array_equal(y, predictions)
-
-
-@pytest.fixture
-def empty_levels():
-    X = [
-        [1],
-        [2],
-        [3],
-    ]
-    y = np.array(
-        [
-            ["1"],
-            ["2", "2.1"],
-            ["3", "3.1", "3.1.2"],
-        ],
-        dtype=object,
-    )
-    return X, y
-
-
-def test_empty_levels(empty_levels):
-    lcppn = LocalClassifierPerParentNode()
-    X, y = empty_levels
-    lcppn.fit(X, y)
-    predictions = lcppn.predict(X)
-    ground_truth = [
-        ["1", "", ""],
-        ["2", "2.1", ""],
-        ["3", "3.1", "3.1.2"],
-    ]
-    assert list(lcppn.hierarchy_.nodes) == [
-        "1",
-        "2",
-        "2" + lcppn.separator_ + "2.1",
-        "3",
-        "3" + lcppn.separator_ + "3.1",
-        "3" + lcppn.separator_ + "3.1" + lcppn.separator_ + "3.1.2",
-        lcppn.root_,
-    ]
-    assert_array_equal(ground_truth, predictions)
-
-
-def test_bert():
-    bert = ConstantClassifier()
-    lcpn = LocalClassifierPerParentNode(
-        local_classifier=bert,
-        bert=True,
-    )
-    X = ["Text 1", "Text 2"]
-    y = ["a", "a"]
-    lcpn.fit(X, y)
-    check_is_fitted(lcpn)
-    predictions = lcpn.predict(X)
-    assert_array_equal(y, predictions)
-
-
-def test_knn():
-    knn = KNeighborsClassifier(
-        n_neighbors=2,
-    )
-    lcppn = LocalClassifierPerParentNode(
-        local_classifier=knn,
-    )
-    y = np.array([["a", "b"], ["a", "c"]])
-    X = np.array([[1, 2], [3, 4]])
-    lcppn.fit(X, y)
-    check_is_fitted(lcppn)
-    # predictions = lcppn.predict(X)
-    # assert_array_equal(y, predictions)
-
-
-def test_fit_multiple_dim_input():
-    lcppn = LocalClassifierPerParentNode()
-    X = np.random.rand(1, 275, 3)
-    y = np.array([["a", "b", "c"]])
-    lcppn.fit(X, y)
-    check_is_fitted(lcppn)
-
-
-def test_predict_multiple_dim_input():
-    lcppn = LocalClassifierPerParentNode()
-    X = np.random.rand(1, 275, 3)
-    y = np.array([["a", "b", "c"]])
-    lcppn.fit(X, y)
-    predictions = lcppn.predict(X)
-    assert predictions is not None
